requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
Flask==2.2.2
h11==0.8.1
h2==3.2.0
hpack==3.0.0
html5lib==1.1
http3==0.6.7
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
requests==2.28.1
requests-async==0.6.2
rfc3986==1.5.0
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.3.0
urllib3==1.26.11
webencodings==0.5.1
Werkzeug==2.2.2
zipp==3.8.1
